backend/requirements/common.txt
django==3.2.9
psycopg2-binary==2.9.2
djangorestframework==3.12.4
django-extensions==3.1.5
django-phonenumber-field==6.0.0
phonenumbers==8.12.3
drf-spectacular==0.21.0
python-dateutil==2.8.2

